+++ packages/gatsby-theme-blog-core/CHANGELOG.md
@@ -3,6 +3,66 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.27...gatsby-theme-blog-core@1.0.28) (2019-11-22)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.26...gatsby-theme-blog-core@1.0.27) (2019-11-20)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.25...gatsby-theme-blog-core@1.0.26) (2019-11-19)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.24...gatsby-theme-blog-core@1.0.25) (2019-11-18)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.24](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.23...gatsby-theme-blog-core@1.0.24) (2019-11-18)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.23](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.22...gatsby-theme-blog-core@1.0.23) (2019-11-18)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.22](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.21...gatsby-theme-blog-core@1.0.22) (2019-11-15)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.21](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.20...gatsby-theme-blog-core@1.0.21) (2019-11-15)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.20](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.19...gatsby-theme-blog-core@1.0.20) (2019-11-15)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.18...gatsby-theme-blog-core@1.0.19) (2019-11-13)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.17...gatsby-theme-blog-core@1.0.18) (2019-11-13)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.16...gatsby-theme-blog-core@1.0.17) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.15...gatsby-theme-blog-core@1.0.16) (2019-11-06)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.14...gatsby-theme-blog-core@1.0.15) (2019-11-05)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
+## [1.0.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.13...gatsby-theme-blog-core@1.0.14) (2019-11-04)
+
+**Note:** Version bump only for package gatsby-theme-blog-core
+
 ## [1.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-theme-blog-core@1.0.12...gatsby-theme-blog-core@1.0.13) (2019-10-29)
 
 **Note:** Version bump only for package gatsby-theme-blog-core